refactor(global): CustomException 패키지 변겅
- exception 폴더 생성
- CustomException 을 exception 폴더 내로 이동
- #21
