+++ src/main/java/com/example/wegather/global/exception/customException/AuthenticationException.java
@@ -1,4 +1,4 @@
-package com.example.wegather.global.customException;
+package com.example.wegather.global.exception.customException;
 
 public class AuthenticationException extends RuntimeException{
