+++ src/main/java/com/example/wegather/global/exception/customException/NoPermissionException.java
@@ -1,4 +1,4 @@
-package com.example.wegather.global.customException;
+package com.example.wegather.global.exception.customException;
 
 public class NoPermissionException extends RuntimeException {
   public NoPermissionException(String message) {
