+++ src/main/java/com/example/wegather/group/domain/service/SmallGroupService.java
@@ -2,7 +2,7 @@ package com.example.wegather.group.domain.service;
 
 import com.example.wegather.auth.MemberDetails;
 import com.example.wegather.global.auth.AuthenticationManagerImpl;
-import com.example.wegather.global.customException.AuthenticationException;
+import com.example.wegather.global.exception.customException.AuthenticationException;
 import com.example.wegather.global.vo.Address;
 import com.example.wegather.group.domain.entity.SmallGroup;
 import com.example.wegather.group.domain.repotitory.SmallGroupRepository;
